requirements.txt
groq>=0.9.0
openai>=1.30.0
pandas>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
statsmodels>=0.14.0
seaborn>=0.13.0
matplotlib>=3.8.0
jupyter>=1.0.0
ipykernel>=6.29.0
